requirements.txt
streamlit==1.32.0
langchain==0.1.11
langchain-openai==0.0.8
langchain-community==0.0.27
langchain-core==0.1.30
langgraph==0.0.27
pandas==2.2.1
requests==2.31.0
gspread==6.0.2
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
openai==1.13.3
python-dotenv==1.0.1
Authlib==1.3.0
streamlit-authenticator==0.3.1
plotly==5.19.0
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
supabase-py==2.3.5
pydantic==2.6.3
sentence-transformers==2.5.1
